实现匹配系统微服务

+++ backedcloud/matchingsystem/src/main/java/com/kob/matchingsystem/service/MatchingService.java
@@ -1,6 +1,6 @@
 package com.kob.matchingsystem.service;
 
 public interface MatchingService {
-    String addPlayer(Long userId, Integer rating);
+    String addPlayer(Long userId, Long rating);
     String removePlayer(Long userId);
 }
